flask-api/requirements.txt
tensorflow==2.8.0
keras==2.8.0
librosa==0.9.1
numpy==1.21.5
pandas
Flask==2.0.3
gunicorn==20.1.0
google-cloud-datastore==2.4.0